refactor(localisation): cleaner and more rusty code, tests also
