+++ src/utils/macros.rs
@@ -9,6 +9,16 @@ macro_rules! command_error {
 }
 
 #[macro_export]
+macro_rules! command_raw_error {
+    ( $arg:expr ) => {
+        $crate::Error::from($arg)
+    };
+    ( $fmt:expr, $( $arg:tt )* ) => {
+        $crate::Error::from(format!($fmt, $($arg)*))
+    };
+}
+
+#[macro_export]
 macro_rules! unwrap_or_return {
     ($_data:expr, $_str:expr) => {
         match $_data {
